Bugfix in paxquery-pact: StringValue vs String
Fixed a bug in which a StringValue object was compared for equality
with a String object. The result returned by StrinValue.equals() would
always be “false” since types are different. This made where statements
return erroneous results when comparing to string literals.

The solution is comparing the output of StringValue.toString() to the
other String object.